+++ modules/sdk/gradle-plugins-soy/src/main/java/com/liferay/gradle/plugins/soy/SoyPlugin.java
@@ -30,16 +30,12 @@ import org.gradle.api.Action;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
-import org.gradle.api.artifacts.Configuration;
-import org.gradle.api.artifacts.DependencySet;
-import org.gradle.api.file.FileCollection;
 import org.gradle.api.file.SourceDirectorySet;
 import org.gradle.api.plugins.BasePlugin;
 import org.gradle.api.plugins.JavaPlugin;
 import org.gradle.api.plugins.PluginContainer;
 import org.gradle.api.tasks.SourceSet;
 import org.gradle.api.tasks.SourceSetOutput;
-import org.gradle.api.tasks.TaskContainer;
 
 /**
  * @author Andrea Di Giorgi
@@ -48,46 +44,13 @@ public class SoyPlugin implements Plugin<Project> {
 
 	public static final String BUILD_SOY_TASK_NAME = "buildSoy";
 
-	public static final String CONFIGURATION_NAME = "soy";
-
 	public static final String WRAP_SOY_ALLOY_TEMPLATE_TASK_NAME =
 		"wrapSoyAlloyTemplate";
 
 	@Override
 	public void apply(Project project) {
-		Configuration soyConfiguration = _addConfigurationSoy(project);
-
 		_addTaskBuildSoy(project);
 		_addTaskWrapSoyAlloyTemplate(project);
-
-		_configureTasksBuildSoy(project, soyConfiguration);
-	}
-
-	private Configuration _addConfigurationSoy(final Project project) {
-		Configuration configuration = GradleUtil.addConfiguration(
-			project, CONFIGURATION_NAME);
-
-		configuration.defaultDependencies(
-			new Action<DependencySet>() {
-
-				@Override
-				public void execute(DependencySet dependencySet) {
-					_addDependenciesSoy(project);
-				}
-
-			});
-
-		configuration.setDescription(
-			"Configures Closure Templates for this project.");
-		configuration.setVisible(false);
-
-		return configuration;
-	}
-
-	private void _addDependenciesSoy(Project project) {
-		GradleUtil.addDependency(
-			project, CONFIGURATION_NAME, "com.google.template", "soy",
-			_VERSION);
 	}
 
 	private BuildSoyTask _addTaskBuildSoy(Project project) {
@@ -172,12 +135,6 @@ public class SoyPlugin implements Plugin<Project> {
 		return wrapSoyAlloyTemplateTask;
 	}
 
-	private void _configureTaskBuildSoyClasspath(
-		BuildSoyTask buildSoyTask, FileCollection fileCollection) {
-
-		buildSoyTask.setClasspath(fileCollection);
-	}
-
 	private void _configureTaskBuildSoyForJavaPlugin(
 		final BuildSoyTask buildSoyTask) {
 
@@ -187,24 +144,6 @@ public class SoyPlugin implements Plugin<Project> {
 				@Override
 				public File call() throws Exception {
 					return _getResourcesDir(buildSoyTask.getProject());
-				}
-
-			});
-	}
-
-	private void _configureTasksBuildSoy(
-		Project project, final Configuration soyConfiguration) {
-
-		TaskContainer taskContainer = project.getTasks();
-
-		taskContainer.withType(
-			BuildSoyTask.class,
-			new Action<BuildSoyTask>() {
-
-				@Override
-				public void execute(BuildSoyTask buildSoyTask) {
-					_configureTaskBuildSoyClasspath(
-						buildSoyTask, soyConfiguration);
 				}
 
 			});
@@ -253,7 +192,5 @@ public class SoyPlugin implements Plugin<Project> {
 
 		return iterator.next();
 	}
-
-	private static final String _VERSION = "2015-04-10";
 
 }