+++ modules/sdk/gradle-plugins-defaults/CHANGELOG.markdown
@@ -1366,11 +1366,19 @@ fix for all tasks of type `FindBugs` related to source sets.
 ### Changed
 - [LPS-71722]: Update the [Liferay Gradle Plugins] dependency to version 3.2.39.
 
-## 3.6.0 - 2017-04-13
+## 3.6.0 - 2017-04-14
 
 ### Added
 - [LPS-71901]: Add task `updateFileSnapshotVersions` to update the project
 version in external files to the latest snapshot.
+
+### Changed
+- [LPS-71164]: Update the [Liferay Gradle Plugins] dependency to version 3.2.40.
+
+## 3.6.1 - 2017-04-14
+
+### Changed
+- [LPS-69139]: Update the [Liferay Gradle Plugins] dependency to version 3.2.41.
 
 [Liferay CDN]: https://cdn.lfrs.sl/repository.liferay.com/nexus/content/groups/public
 [Liferay Gradle Plugins]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins
@@ -1442,6 +1450,7 @@ version in external files to the latest snapshot.
 [LPS-68980]: https://issues.liferay.com/browse/LPS-68980
 [LPS-69013]: https://issues.liferay.com/browse/LPS-69013
 [LPS-69026]: https://issues.liferay.com/browse/LPS-69026
+[LPS-69139]: https://issues.liferay.com/browse/LPS-69139
 [LPS-69223]: https://issues.liferay.com/browse/LPS-69223
 [LPS-69271]: https://issues.liferay.com/browse/LPS-69271
 [LPS-69288]: https://issues.liferay.com/browse/LPS-69288